delete button is working now

+++ public/js/deletePost.js
@@ -1,26 +1,22 @@
 async function deleteFormHandler(event) {
-    event.preventDefault();
+  event.preventDefault();
+  
+  if (event.target.tagName !== 'BUTTON') {
+    return;
+  }
+  const id = event.target.getAttribute('data-id')
 
-    const id = window.location.toString().split('/')[
-        window.location.toString().split('/').length - 1
-      ];
+  const response = await fetch(`/api/posts/${id}`, {
+    method: 'DELETE',
 
-    const response = await fetch(`/api/posts/${id}`, {
-        method: 'DELETE',
-        body: JSON.stringify({
-          post_id: id
-        }),
-        headers: {
-          'Content-Type': 'application/json'
-        }
-      });
+  });
 
-      if (response.ok) {
-        document.location.replace('/dashboard/');
-      } else {
-        alert(response.statusText);
-      }
-
+  if (response.ok) {
+    document.location.replace('/dashboard/');
+  } else {
+    alert(response.statusText);
   }
 
-  document.querySelector('.delete-post-btn').addEventListener('click', deleteFormHandler);
+}
+
+document.querySelector('.project-list').addEventListener('click', deleteFormHandler);
